refact: move ember-try scenario into test-app
The test-app is the one that we will test with ember-try for now.
We want to make sure the addon works when included into certain
versions of host ember apps. This is better reflected when moving
the try scenario into the test-app than to run ember-try on the
addon test-suite instead.