docs: deploy versioned docs
Mimics ember-cli-addon-docs (ecad) behavior and deploys the
documentation into

    versions/_VERSION_

where the _VERSION_ can either be the tag or the branch. The
versions.json in the root of the deployed docs is updated accordingly
and follows the format of ecad so it is backwards compatible to versions
of the ember-statecharts documentation which used ecad.

In case a documentation should be used as the current (latest) one, we
can deploy it to the root of the deployed docs. The versions.json is
updated accordingly as well.

---

There are 3 deploy targets which can be used:

    ember deploy tag
    ember deploy branch
    ember deploy latest

- `tag` will use the tag as the name for the deployed version

- `branch` will use the branch for the deployed version; this deploy
  target can be used to continuously deploy a branch, say `beta` or
  `master`

Those two deploy into versions/${tag} or versions/${branch}

- `latest` will clear out the deployed docs and replace the content with
  the latest documentation, it will also restore the content of
  `versions/` and versions.json